refactor: seperate collector and exporter
modify collector struct - only collect kafka connect statuses
add exporter struct - export prometheus metrics
