+++ internal/exporter/prometheus/types.go
@@ -1,0 +1,10 @@
+package exporter
+
+type ConnectorStatusMetric struct {
+	Name                string
+	UnAssignedTaskCount int
+	RunningTaskCount    int
+	PausedTaskCount     int
+	FailedTaskCount     int
+	TotalTaskCount      int
+}
